Reduce the default timeout to 1m
This is a bit of a tricky one. We need to balance potentially having
enough time to pull a couple of Functions with not making folks wait
forever for the context to timeout if something goes wrong (e.g. a
Function fails to run and we're waiting for our gRPC call to timeout).

Signed-off-by: Nic Cope <nicc@rk0n.org>