Load all observed resources
I thought loading observed resources was completely broken, but it's
not (there was an issue with my YAML). In my quest to figure this out I
added a test, and refactored slightly the code that loads observed
resources to support multiple files/directories of observed resources (I
think that was the original intent).

Signed-off-by: Nic Cope <nicc@rk0n.org>